Password hashing with BCrypt encoder
Set a password hashing mechanism with BCrypt algorithm, so there is no plain password now in user records of the database.

Root user is now being generated in memory with a random password each time is run the application. No longer can be persisted to memory, and by this cannot edit profile or change password.